Remove redundant comments, assertions, organize better
