+++ src/multithreading/create-worker.test.ts
@@ -7,18 +7,22 @@ import { createWorker } from './create-worker';
 
 describe(createWorker.name, () => {
   it('should create a Worker instance when valid parameters are provided', async () => {
+    // Arrange
     const workerPath = __dirname + '../tests/dummy-worker.ts';
+    const event = createMockEvent(mockServer.baseUrl, {
+      payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
+    });
 
+    // Act
     const worker = isMainThread
       ? await createWorker<object>({
-          event: createMockEvent(mockServer.baseUrl, {
-            payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
-          }),
+          event,
           initialState: {},
           workerPath,
         })
       : null;
 
+    // Assert
     expect(worker).not.toBeNull();
     expect(worker).toBeInstanceOf(Worker);
 
@@ -28,44 +32,51 @@ describe(createWorker.name, () => {
   });
 
   it('should throw error when not in main thread', async () => {
+    // Arrange
     const originalIsMainThread = isMainThread;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     (isMainThread as any) = false;
     const workerPath = __dirname + '../tests/dummy-worker.ts';
+    const event = createMockEvent(mockServer.baseUrl, {
+      payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
+    });
 
+    // Act & Assert
     await expect(
       createWorker<object>({
-        event: createMockEvent(mockServer.baseUrl, {
-          payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
-        }),
+        event,
         initialState: {},
         workerPath,
       })
     ).rejects.toThrow('Worker threads can not start more worker threads.');
 
-    // Restore original value
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     (isMainThread as any) = originalIsMainThread;
   });
 
   it('[edge] should handle worker creation with minimal valid data', async () => {
+    // Arrange
     const workerPath = __dirname + '../tests/dummy-worker.ts';
+    const event = createMockEvent(mockServer.baseUrl, {
+      payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
+    });
 
     if (isMainThread) {
+      // Act
       const worker = await createWorker<object>({
-        event: createMockEvent(mockServer.baseUrl, {
-          payload: { event_type: EventType.ExtractionExternalSyncUnitsStart },
-        }),
+        event,
         initialState: {},
         workerPath,
       });
 
+      // Assert
       expect(worker).toBeInstanceOf(Worker);
       await worker.terminate();
     }
   });
 
   it('[edge] should handle worker creation with complex initial state', async () => {
+    // Arrange
     const workerPath = __dirname + '../tests/dummy-worker.ts';
     const complexState = {
       nested: {
@@ -73,33 +84,40 @@ describe(createWorker.name, () => {
         config: { enabled: true },
       },
     };
+    const event = createMockEvent(mockServer.baseUrl, {
+      payload: { event_type: EventType.ExtractionDataStart },
+    });
 
     if (isMainThread) {
+      // Act
       const worker = await createWorker<typeof complexState>({
-        event: createMockEvent(mockServer.baseUrl, {
-          payload: { event_type: EventType.ExtractionDataStart },
-        }),
+        event,
         initialState: complexState,
         workerPath,
       });
 
+      // Assert
       expect(worker).toBeInstanceOf(Worker);
       await worker.terminate();
     }
   });
 
   it('[edge] should handle different event types', async () => {
+    // Arrange
     const workerPath = __dirname + '../tests/dummy-worker.ts';
+    const event = createMockEvent(mockServer.baseUrl, {
+      payload: { event_type: EventType.ExtractionMetadataStart },
+    });
 
     if (isMainThread) {
+      // Act
       const worker = await createWorker<object>({
-        event: createMockEvent(mockServer.baseUrl, {
-          payload: { event_type: EventType.ExtractionMetadataStart },
-        }),
+        event,
         initialState: {},
         workerPath,
       });
 
+      // Assert
       expect(worker).toBeInstanceOf(Worker);
       await worker.terminate();
     }
