+++ src/multithreading/worker-adapter/worker-adapter.loading.test.ts
@@ -111,34 +111,34 @@ describe(`${WorkerAdapter.name}.loadItemTypes — timeout and unexpected errors`
   });
 
   it('should emit DataLoadingProgress and exit on timeout', async () => {
+    // Arrange
     const items = [makeLoaderItem('dev-1'), makeLoaderItem('dev-2')];
     setupLoaderFile(adapter, items);
-
     adapter.isTimeout = true;
-
     const itemTypesToLoad = [
       { itemType: 'tasks', create: jest.fn(), update: jest.fn() },
     ];
 
+    // Act
     await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(LoaderEventType.DataLoadingProgress);
     expect(exitSpy).toHaveBeenCalledWith(0);
   });
 
   it('should emit DataLoadingProgress mid-loop when timeout arrives between items', async () => {
+    // Arrange
     const items = [
       makeLoaderItem('dev-1'),
       makeLoaderItem('dev-2'),
       makeLoaderItem('dev-3'),
     ];
     setupLoaderFile(adapter, items);
-
     exitSpy.mockRestore();
     exitSpy = jest.spyOn(process, 'exit').mockImplementation((() => {
       throw new Error('process.exit');
     }) as never);
-
     let loadItemCallCount = 0;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/require-await
     jest.spyOn(adapter as any, 'loadItem').mockImplementation(async () => {
@@ -148,20 +148,20 @@ describe(`${WorkerAdapter.name}.loadItemTypes — timeout and unexpected errors`
       }
       return { report: { item_type: 'tasks', updated: 1 } };
     });
-
     const itemTypesToLoad = [
       { itemType: 'tasks', create: jest.fn(), update: jest.fn() },
     ];
 
+    // Act & Assert
     await expect(adapter.loadItemTypes({ itemTypesToLoad })).rejects.toThrow(
       'process.exit'
     );
-
     expect(loadItemCallCount).toBe(1);
     expect(emitSpy).toHaveBeenCalledWith(LoaderEventType.DataLoadingProgress);
   });
 
   it('should emit DataLoadingError and exit(1) on unexpected error', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = {
       filesToLoad: [
         {
@@ -174,17 +174,17 @@ describe(`${WorkerAdapter.name}.loadItemTypes — timeout and unexpected errors`
         },
       ],
     };
-
     adapter['uploader'].getJsonObjectByArtifactId = jest
       .fn()
       .mockRejectedValue(new Error('Unexpected network failure'));
-
     const itemTypesToLoad = [
       { itemType: 'tasks', create: jest.fn(), update: jest.fn() },
     ];
 
+    // Act
     await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.DataLoadingError,
       expect.objectContaining({
@@ -227,6 +227,7 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
   });
 
   it('should accumulate an UPDATED report when the connector updates the item and the mapper sync succeeds', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-1')]);
     adapter['_mappers'].getByTargetId = jest.fn().mockResolvedValue({
       data: { sync_mapper_record: { id: 'smr-1' } },
@@ -236,8 +237,10 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
       .fn()
       .mockResolvedValue({ id: 'ext-updated-1' });
 
+    // Act
     const { reports } = await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(reports).toEqual(
       expect.arrayContaining([
         expect.objectContaining({
@@ -250,6 +253,7 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
   });
 
   it('should fall back to create and accumulate a CREATED report when the mapper record does not exist (404)', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-2')]);
     const axiosError = { isAxiosError: true, response: { status: 404 } };
     adapter['_mappers'].getByTargetId = jest.fn().mockRejectedValue(axiosError);
@@ -258,8 +262,10 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
       .fn()
       .mockResolvedValue({ id: 'new-ext-id' });
 
+    // Act
     const { reports } = await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(reports).toEqual(
       expect.arrayContaining([
         expect.objectContaining({
@@ -272,14 +278,17 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
   });
 
   it('should emit DataLoadingDelayed and stop processing when the connector signals a rate-limit delay', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-3')]);
     adapter['_mappers'].getByTargetId = jest.fn().mockResolvedValue({
       data: { sync_mapper_record: { id: 'smr-1' } },
     });
     itemTypesToLoad[0].update = jest.fn().mockResolvedValue({ delay: 15 });
 
+    // Act
     await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.DataLoadingDelayed,
       expect.objectContaining({ delay: 15 })
@@ -287,6 +296,7 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
   });
 
   it('should count the item as FAILED when the update succeeds but the mapper sync throws', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-4')]);
     adapter['_mappers'].getByTargetId = jest.fn().mockResolvedValue({
       data: { sync_mapper_record: { id: 'smr-1' } },
@@ -296,12 +306,16 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
       .mockRejectedValue(new Error('mapper down'));
     itemTypesToLoad[0].update = jest.fn().mockResolvedValue({ id: 'ext-id' });
 
+    // Act
     const { reports } = await adapter.loadItemTypes({ itemTypesToLoad });
+
+    // Assert
     expect(emitSpy).not.toHaveBeenCalled();
     expect(reports).toBeDefined();
   });
 
   it('should not emit for a non-404 Axios error from the mapper (recorded as item-level error)', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-5')]);
     const axiosError = {
       isAxiosError: true,
@@ -310,18 +324,24 @@ describe(`${WorkerAdapter.name}.loadItemTypes — loadItem branch coverage via p
     };
     adapter['_mappers'].getByTargetId = jest.fn().mockRejectedValue(axiosError);
 
+    // Act
     await adapter.loadItemTypes({ itemTypesToLoad });
+
+    // Assert
     expect(emitSpy).not.toHaveBeenCalled();
   });
 
   it('should handle a null sync_mapper_record gracefully and continue loading', async () => {
+    // Arrange
     setupLoaderFile(adapter, [makeLoaderItem('dev-6')]);
     adapter['_mappers'].getByTargetId = jest
       .fn()
       .mockResolvedValue({ data: null });
 
+    // Act
     const { reports } = await adapter.loadItemTypes({ itemTypesToLoad });
 
+    // Assert
     expect(emitSpy).not.toHaveBeenCalled();
     expect(reports).toBeDefined();
   });
@@ -345,19 +365,23 @@ describe(`${WorkerAdapter.name}.loadItemTypes — additional branches`, () => {
   });
 
   it('should return immediately with empty reports when filesToLoad is empty', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = { filesToLoad: [] };
 
+    // Act
     const result = await adapter.loadItemTypes({
       itemTypesToLoad: [
         { itemType: 'tasks', create: jest.fn(), update: jest.fn() },
       ],
     });
 
+    // Assert
     expect(result.reports).toEqual([]);
     expect(emitSpy).not.toHaveBeenCalled();
   });
 
   it('should emit DataLoadingError when a file references an item type not in itemTypesToLoad', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = {
       filesToLoad: [
         {
@@ -374,12 +398,14 @@ describe(`${WorkerAdapter.name}.loadItemTypes — additional branches`, () => {
       .fn()
       .mockResolvedValue({ response: [makeLoaderItem()] });
 
+    // Act
     await adapter.loadItemTypes({
       itemTypesToLoad: [
         { itemType: 'tasks', create: jest.fn(), update: jest.fn() },
       ],
     });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.DataLoadingError,
       expect.objectContaining({
@@ -433,6 +459,7 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
   });
 
   it('should emit AttachmentLoadingProgress and exit on timeout', async () => {
+    // Arrange
     const items = [
       {
         reference_id: 'ref-1',
@@ -450,13 +477,14 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
       },
     ] as ExternalSystemAttachment[];
     setupFilesToLoad(adapter, items);
-
     adapter.isTimeout = true;
 
+    // Act
     await adapter.loadAttachments({
       create: jest.fn(),
     });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.AttachmentLoadingProgress
     );
@@ -464,6 +492,7 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
   });
 
   it('should emit AttachmentLoadingError on transformer file error', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = {
       filesToLoad: [
         {
@@ -476,7 +505,6 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
         },
       ],
     };
-
     adapter['uploader'].getJsonObjectByArtifactId = jest
       .fn()
       .mockResolvedValue({
@@ -484,10 +512,12 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
         error: new Error('Artifact not found'),
       });
 
+    // Act
     await adapter.loadAttachments({
       create: jest.fn(),
     });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.AttachmentLoadingError,
       expect.objectContaining({
@@ -499,6 +529,7 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
   });
 
   it('should emit AttachmentLoadingError and exit(1) on unexpected error', async () => {
+    // Arrange
     const items = [
       {
         reference_id: 'ref-1',
@@ -516,13 +547,14 @@ describe(`${WorkerAdapter.name}.loadAttachments — timeout, transformer errors,
       },
     ] as ExternalSystemAttachment[];
     setupFilesToLoad(adapter, items);
-
     const mockCreate = jest
       .fn()
       .mockRejectedValue(new Error('Unexpected API failure'));
 
+    // Act
     await adapter.loadAttachments({ create: mockCreate });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.AttachmentLoadingError,
       expect.objectContaining({
@@ -553,15 +585,19 @@ describe(`${WorkerAdapter.name}.loadAttachments — additional branches`, () => 
   });
 
   it('should return immediately with empty reports when fromDevRev is not set', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = undefined;
 
+    // Act
     const result = await adapter.loadAttachments({ create: jest.fn() });
 
+    // Assert
     expect(result.reports).toEqual([]);
     expect(emitSpy).not.toHaveBeenCalled();
   });
 
   it('should emit AttachmentLoadingDelayed and stop the loop when the connector signals a rate-limit delay', async () => {
+    // Arrange
     adapter['adapterState'].state.fromDevRev = {
       filesToLoad: [
         {
@@ -594,14 +630,15 @@ describe(`${WorkerAdapter.name}.loadAttachments — additional branches`, () => 
           },
         ],
       });
-
     jest
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       .spyOn(adapter as any, 'loadAttachment')
       .mockResolvedValue({ rateLimit: { delay: 20 } } as never);
 
+    // Act
     await adapter.loadAttachments({ create: jest.fn() });
 
+    // Assert
     expect(emitSpy).toHaveBeenCalledWith(
       LoaderEventType.AttachmentLoadingDelayed,
       expect.objectContaining({ delay: 20 })
@@ -642,51 +679,63 @@ describe(`${WorkerAdapter.name}.loadAttachment`, () => {
   }
 
   it('should return a CREATED report when create succeeds', async () => {
+    // Arrange
     adapter['_mappers'].create = jest.fn().mockResolvedValue({ data: {} });
     const create = jest.fn().mockResolvedValue({ id: 'att-ext-1' });
 
+    // Act
     const result = await adapter['loadAttachment']({
       item: makeAttachment(),
       create,
     });
 
+    // Assert
     expect(result.report?.item_type).toBe('attachment');
     expect(result.report?.[ActionType.CREATED]).toBe(1);
   });
 
   it('should still return CREATED even when mapper create fails — attachment loading is resilient', async () => {
+    // Arrange
     adapter['_mappers'].create = jest
       .fn()
       .mockRejectedValue(new Error('mapper failed'));
     const create = jest.fn().mockResolvedValue({ id: 'att-ext-1' });
 
+    // Act
     const result = await adapter['loadAttachment']({
       item: makeAttachment(),
       create,
     });
 
+    // Assert
     expect(result.report?.[ActionType.CREATED]).toBe(1);
   });
 
   it('should propagate rate-limit delay when the connector signals one', async () => {
+    // Arrange
     const create = jest.fn().mockResolvedValue({ delay: 30 });
 
+    // Act
     const result = await adapter['loadAttachment']({
       item: makeAttachment(),
       create,
     });
 
+    // Assert
     expect(result.rateLimit?.delay).toBe(30);
   });
 
   it('should return a FAILED report when create returns neither id nor delay', async () => {
+    // Arrange
     const create = jest.fn().mockResolvedValue({ id: null, delay: null });
 
+    // Act
     const result = await adapter['loadAttachment']({
       item: makeAttachment(),
       create,
     });
 
+    // Assert
     expect(result.report?.[ActionType.FAILED]).toBe(1);
   });
 });
