+++ src/repo/repo.test.ts
@@ -31,15 +31,20 @@ describe(Repo.name, () => {
   });
 
   it('should normalize and push items when array contains items', async () => {
+    // Arrange
     const items = createItems(10);
-    await repo.push(items);
-    expect(normalize).toHaveBeenCalledTimes(10);
 
+    // Act
+    await repo.push(items);
+
+    // Assert
+    expect(normalize).toHaveBeenCalledTimes(10);
     const normalizedItems = items.map((item) => normalizeItem(item));
     expect(repo.getItems()).toEqual(normalizedItems);
   });
 
   it('should not normalize items when normalize function is not provided', async () => {
+    // Arrange
     repo = new Repo({
       event: createMockEvent(mockServer.baseUrl, {
         payload: { event_type: EventType.ExtractionDataStart },
@@ -48,18 +53,25 @@ describe(Repo.name, () => {
       onUpload: jest.fn(),
       options: {},
     });
-
     const items = createItems(10);
+
+    // Act
     await repo.push(items);
+
+    // Assert
     expect(normalize).not.toHaveBeenCalled();
   });
 
   it('[edge] should not push items when items array is empty', async () => {
+    // Act
     await repo.push([]);
+
+    // Assert
     expect(repo.getItems()).toEqual([]);
   });
 
   it('should not normalize items when item type is external_domain_metadata', async () => {
+    // Arrange
     repo = new Repo({
       event: createMockEvent(mockServer.baseUrl, {
         payload: { event_type: EventType.ExtractionDataStart },
@@ -69,14 +81,17 @@ describe(Repo.name, () => {
       onUpload: jest.fn(),
       options: {},
     });
-
     const items = createItems(10);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(normalize).not.toHaveBeenCalled();
   });
 
   it('should not normalize items when item type is ssor_attachment', async () => {
+    // Arrange
     repo = new Repo({
       event: createMockEvent(mockServer.baseUrl, {
         payload: { event_type: EventType.ExtractionDataStart },
@@ -86,41 +101,58 @@ describe(Repo.name, () => {
       onUpload: jest.fn(),
       options: {},
     });
-
     const items = createItems(10);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(normalize).not.toHaveBeenCalled();
   });
 
   it('should leave 5 items in the items array after pushing 2005 items with batch size of 2000', async () => {
+    // Arrange
     const items = createItems(2005);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(repo.getItems().length).toBe(5);
   });
 
   it('should normalize all items when pushing 4005 items with batch size of 2000', async () => {
+    // Arrange
     const items = createItems(4005);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(normalize).toHaveBeenCalledTimes(4005);
   });
 
   it('should upload 2 batches when pushing 4005 items with batch size of 2000', async () => {
+    // Arrange
     const uploadSpy = jest.spyOn(repo, 'upload');
-
     const items = createItems(4005);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(uploadSpy).toHaveBeenCalledTimes(2);
     uploadSpy.mockRestore();
   });
 
   it('should leave 5 items in array after pushing 4005 items with batch size of 2000', async () => {
+    // Arrange
     const items = createItems(4005);
+
+    // Act
     await repo.push(items);
 
+    // Assert
     expect(repo.getItems().length).toBe(5);
   });
 
@@ -140,39 +172,59 @@ describe(Repo.name, () => {
     });
 
     it('should empty the items array after pushing 50 items with batch size of 50', async () => {
+      // Arrange
       const items = createItems(50);
+
+      // Act
       await repo.push(items);
+
+      // Assert
       expect(repo.getItems()).toEqual([]);
     });
 
     it('should leave 5 items in the items array after pushing 205 items with batch size of 50', async () => {
+      // Arrange
       const items = createItems(205);
+
+      // Act
       await repo.push(items);
 
+      // Assert
       expect(repo.getItems().length).toBe(5);
     });
 
     it('should normalize all items when pushing 205 items with batch size of 50', async () => {
+      // Arrange
       const items = createItems(205);
+
+      // Act
       await repo.push(items);
 
+      // Assert
       expect(normalize).toHaveBeenCalledTimes(205);
     });
 
     it('should upload 4 batches when pushing 205 items with batch size of 50', async () => {
+      // Arrange
       const uploadSpy = jest.spyOn(repo, 'upload');
-
       const items = createItems(205);
+
+      // Act
       await repo.push(items);
 
+      // Assert
       expect(uploadSpy).toHaveBeenCalledTimes(4);
       uploadSpy.mockRestore();
     });
 
     it('should leave 5 items in array after pushing 205 items with batch size of 50', async () => {
+      // Arrange
       const items = createItems(205);
+
+      // Act
       await repo.push(items);
 
+      // Assert
       expect(repo.getItems().length).toBe(5);
     });
   });
