+++ src/state/state.test.ts
@@ -4,7 +4,7 @@ import {
 } from '../common/constants';
 import { mockServer } from '../tests/jest.setup';
 import { createMockEvent } from '../common/test-utils';
-import { EventType, TimeValue, TimeValueType } from '../types/extraction';
+import { EventType } from '../types/extraction';
 import { State, createAdapterState } from './state';
 import { extractionSdkState } from './state.interfaces';
 
@@ -300,574 +300,6 @@ describe(State.name, () => {
     }
   );
 
-  describe('Enhanced Control Protocol - TimeValue resolution failures', () => {
-    it('should exit the process if extraction_start_time resolution fails', async () => {
-      // Arrange: WORKERS_NEWEST type but state has no workersNewest
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.WORKERS_NEWEST,
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-        workers_oldest: '',
-        workers_newest: '',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-      jest.spyOn(console, 'error').mockImplementation(() => {});
-
-      // Act & Assert
-      await expect(
-        createAdapterState({
-          event,
-          initialState: {},
-          initialDomainMapping: {},
-        })
-      ).rejects.toThrow('process.exit called');
-      expect(processExitSpy).toHaveBeenCalledWith(1);
-    });
-
-    it('should exit the process if extraction_end_time resolution fails', async () => {
-      // Arrange: WORKERS_NEWEST type but state has no workersNewest
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.UNBOUNDED,
-            },
-            extraction_end_time: {
-              type: TimeValueType.WORKERS_NEWEST,
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-        workers_oldest: '',
-        workers_newest: '',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-      jest.spyOn(console, 'error').mockImplementation(() => {});
-
-      // Act & Assert
-      await expect(
-        createAdapterState({
-          event,
-          initialState: {},
-          initialDomainMapping: {},
-        })
-      ).rejects.toThrow('process.exit called');
-      expect(processExitSpy).toHaveBeenCalledWith(1);
-    });
-  });
-
-  describe('Backwards compatibility - missing TimeValue type', () => {
-    it('should skip resolution when extraction_start_time has no type', async () => {
-      // Arrange: platform sends extraction_start_time without a type field (old platform version)
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {} as unknown as TimeValue,
-            extraction_end_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2025-06-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      const state = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: should not crash, extract_from is not set, extract_to is resolved
-      expect(processExitSpy).not.toHaveBeenCalled();
-      expect(event.payload.event_context.extract_from).toBeUndefined();
-      expect(event.payload.event_context.extract_to).toBe(
-        '2025-06-01T00:00:00.000Z'
-      );
-      expect(state.state.pendingWorkersNewest).toBe('2025-06-01T00:00:00.000Z');
-    });
-
-    it('should skip resolution when extraction_end_time has no type', async () => {
-      // Arrange: platform sends extraction_end_time without a type field
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-01-01T00:00:00Z',
-            },
-            extraction_end_time: {} as unknown as TimeValue,
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: should not crash, extract_to is not set, extract_from is resolved
-      expect(processExitSpy).not.toHaveBeenCalled();
-      expect(event.payload.event_context.extract_from).toBe(
-        '2024-01-01T00:00:00.000Z'
-      );
-      expect(event.payload.event_context.extract_to).toBeUndefined();
-    });
-
-    it('should skip resolution when both extraction times have no type', async () => {
-      // Arrange: platform sends both time values without type fields
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              value: 'some-value',
-            } as unknown as TimeValue,
-            extraction_end_time: {
-              value: 'some-value',
-            } as unknown as TimeValue,
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: should not crash, neither extraction time is resolved
-      expect(processExitSpy).not.toHaveBeenCalled();
-      expect(event.payload.event_context.extract_from).toBeUndefined();
-      expect(event.payload.event_context.extract_to).toBeUndefined();
-    });
-  });
-
-  describe('Enhanced Control Protocol - extraction window validation', () => {
-    it('should exit the process if extract_from >= extract_to', async () => {
-      // Arrange: start is after end (inverted window)
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2025-06-01T00:00:00Z',
-            },
-            extraction_end_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-01-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-      jest.spyOn(console, 'error').mockImplementation(() => {});
-
-      // Act & Assert
-      await expect(
-        createAdapterState({
-          event,
-          initialState: {},
-          initialDomainMapping: {},
-        })
-      ).rejects.toThrow('process.exit called');
-      expect(processExitSpy).toHaveBeenCalledWith(1);
-    });
-
-    it('should exit the process if extract_from equals extract_to', async () => {
-      // Arrange: start equals end (zero-width window)
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-06-01T00:00:00Z',
-            },
-            extraction_end_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-06-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-      jest.spyOn(console, 'error').mockImplementation(() => {});
-
-      // Act & Assert
-      await expect(
-        createAdapterState({
-          event,
-          initialState: {},
-          initialDomainMapping: {},
-        })
-      ).rejects.toThrow('process.exit called');
-      expect(processExitSpy).toHaveBeenCalledWith(1);
-    });
-
-    it('should not exit when extract_from < extract_to', async () => {
-      // Arrange: valid window
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-01-01T00:00:00Z',
-            },
-            extraction_end_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2025-06-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: process.exit should NOT have been called
-      expect(processExitSpy).not.toHaveBeenCalled();
-    });
-
-    it('should not validate when only extract_from is set', async () => {
-      // Arrange: only start, no end
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2024-01-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: process.exit should NOT have been called
-      expect(processExitSpy).not.toHaveBeenCalled();
-    });
-
-    it('should not exit when extract_from is UNBOUNDED and extract_to is a real timestamp', async () => {
-      // Arrange: UNBOUNDED start (epoch) with a real ABSOLUTE end timestamp
-      const event = createMockEvent(mockServer.baseUrl, {
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.UNBOUNDED,
-            },
-            extraction_end_time: {
-              type: TimeValueType.ABSOLUTE_TIME,
-              value: '2025-06-01T00:00:00Z',
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: process.exit should NOT have been called
-      expect(processExitSpy).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Pending extraction boundaries (pendingWorkersOldest/pendingWorkersNewest)', () => {
-    const FIXED_NOW = '2026-03-26T10:00:00.000Z';
-
-    beforeEach(() => {
-      jest.useFakeTimers();
-      jest.setSystemTime(new Date(FIXED_NOW));
-    });
-
-    afterEach(() => {
-      jest.useRealTimers();
-    });
-
-    it('should store resolved values in pendingWorkersOldest/pendingWorkersNewest on StartExtractingMetadata', async () => {
-      // Arrange
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: '',
-        },
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.UNBOUNDED,
-            },
-            extraction_end_time: {
-              type: TimeValueType.CURRENT_TIME,
-            },
-          },
-        },
-      });
-
-      fetchStateSpy.mockRejectedValue({
-        isAxiosError: true,
-        response: { status: 404 },
-      });
-      installInitialDomainMappingSpy.mockResolvedValue({ success: true });
-      postStateSpy.mockResolvedValue({ success: true });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      const state = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert
-      expect(state.state.pendingWorkersOldest).toBe('1970-01-01T00:00:00.000Z');
-      expect(state.state.pendingWorkersNewest).toBe(FIXED_NOW);
-      expect(event.payload.event_context.extract_from).toBe(
-        '1970-01-01T00:00:00.000Z'
-      );
-      expect(event.payload.event_context.extract_to).toBe(FIXED_NOW);
-    });
-
-    it('should overwrite pending values on a retry (new StartExtractingMetadata after failure)', async () => {
-      // Arrange: state has stale pending values from a previous failed attempt
-      const staleOldest = '2026-03-25T08:00:00.000Z';
-      const staleNewest = '2026-03-25T09:00:00.000Z';
-
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: {
-          event_type: EventType.StartExtractingMetadata,
-          event_context: {
-            extraction_start_time: {
-              type: TimeValueType.UNBOUNDED,
-            },
-            extraction_end_time: {
-              type: TimeValueType.CURRENT_TIME,
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-        pendingWorkersOldest: staleOldest,
-        pendingWorkersNewest: staleNewest,
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      const state = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: pending values are overwritten with fresh resolution, not stale values
-      expect(state.state.pendingWorkersOldest).toBe('1970-01-01T00:00:00.000Z');
-      expect(state.state.pendingWorkersNewest).toBe(FIXED_NOW);
-      expect(state.state.pendingWorkersNewest).not.toBe(staleNewest);
-    });
-
-    it('should reuse pending values from state on ContinueExtractingData instead of re-resolving', async () => {
-      // Arrange: state has pending values from a prior StartExtractingMetadata phase
-      const pendingOldest = '1970-01-01T00:00:00.000Z';
-      const pendingNewest = '2026-03-26T08:00:00.000Z'; // Earlier than FIXED_NOW
-
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: {
-          event_type: EventType.ContinueExtractingData,
-          event_context: {
-            // Platform still sends TimeValue objects, but they should be ignored
-            extraction_start_time: {
-              type: TimeValueType.CURRENT_TIME,
-            },
-            extraction_end_time: {
-              type: TimeValueType.CURRENT_TIME,
-            },
-          },
-        },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-        pendingWorkersOldest: pendingOldest,
-        pendingWorkersNewest: pendingNewest,
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      const state = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: uses cached pending values, NOT new Date() resolution
-      expect(event.payload.event_context.extract_from).toBe(pendingOldest);
-      expect(event.payload.event_context.extract_to).toBe(pendingNewest);
-      // Pending values in state remain unchanged
-      expect(state.state.pendingWorkersOldest).toBe(pendingOldest);
-      expect(state.state.pendingWorkersNewest).toBe(pendingNewest);
-    });
-
-    it('should not set extract_from/extract_to on ContinueExtractingData if no pending values exist', async () => {
-      // Arrange: state has no pending values (e.g. old state from before this feature)
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: { event_type: EventType.ContinueExtractingData },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: no extraction timestamps are set
-      expect(event.payload.event_context.extract_from).toBeUndefined();
-      expect(event.payload.event_context.extract_to).toBeUndefined();
-    });
-
-    it('should reuse pending values on StartExtractingAttachments', async () => {
-      // Arrange: state has pending values from the StartExtractingMetadata phase
-      const pendingOldest = '1970-01-01T00:00:00.000Z';
-      const pendingNewest = '2026-03-26T08:00:00.000Z';
-
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: 'test_snap_in_version_id',
-        },
-        payload: { event_type: EventType.StartExtractingAttachments },
-      });
-
-      const stringifiedState = JSON.stringify({
-        snapInVersionId: 'test_snap_in_version_id',
-        pendingWorkersOldest: pendingOldest,
-        pendingWorkersNewest: pendingNewest,
-      });
-      fetchStateSpy.mockResolvedValue({ state: stringifiedState });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      // Act
-      await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Assert: pending values are reused
-      expect(event.payload.event_context.extract_from).toBe(pendingOldest);
-      expect(event.payload.event_context.extract_to).toBe(pendingNewest);
-    });
-  });
-
   it('should populate extractionScope from API response', async () => {
     // Arrange
     const event = createMockEvent(mockServer.baseUrl, {
@@ -969,82 +401,5 @@ describe(State.name, () => {
     );
     expect(result.extractionScope).toEqual({});
     expect(processExitSpy).not.toHaveBeenCalled();
-  });
-
-  describe('State.postState', () => {
-    it('should POST the stringified state with Authorization header to the update endpoint', async () => {
-      // Arrange
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: {
-          snap_in_version_id: '1.0.0',
-          secrets: { service_account_token: 'test_token' },
-        },
-        payload: { event_type: EventType.StartExtractingData },
-      });
-      const stateToPost = { snapInVersionId: '1.0.0', foo: 'bar' };
-      fetchStateSpy.mockResolvedValue({
-        state: JSON.stringify({ snapInVersionId: '1.0.0' }),
-      });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-
-      postStateSpy.mockRestore();
-
-      const adapterState = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      await adapterState.postState(stateToPost as never);
-
-      // The mock server records all incoming requests — inspect what was sent
-      const requests = mockServer.getRequests(
-        'POST',
-        '/worker_data_url.update'
-      );
-      expect(requests).toHaveLength(1);
-
-      const body = requests[0].body as { state: string };
-      // Body must contain the stringified state, preserving the original fields
-      expect(typeof body.state).toBe('string');
-      const parsed = JSON.parse(body.state) as Record<string, unknown>;
-      expect(parsed.foo).toBe('bar');
-      expect(parsed.snapInVersionId).toBe('1.0.0');
-    });
-
-    it('should exit(1) when postState HTTP request fails', async () => {
-      // Arrange
-      const event = createMockEvent(mockServer.baseUrl, {
-        context: { snap_in_version_id: '1.0.0' },
-        payload: { event_type: EventType.StartExtractingData },
-      });
-      fetchStateSpy.mockResolvedValue({
-        state: JSON.stringify({ snapInVersionId: '1.0.0' }),
-      });
-      jest.spyOn(console, 'log').mockImplementation(() => {});
-      jest.spyOn(console, 'error').mockImplementation(() => {});
-
-      postStateSpy.mockRestore();
-
-      const adapterState = await createAdapterState({
-        event,
-        initialState: {},
-        initialDomainMapping: {},
-      });
-
-      // Mock axiosClient.post directly to bypass the retry backoff
-
-      const axiosClientModule = require('../http/axios-client-internal');
-      const axiosPostSpy = jest
-        .spyOn(axiosClientModule.axiosClient, 'post')
-        .mockRejectedValue(new Error('network error'));
-
-      await expect(adapterState.postState()).rejects.toThrow(
-        'process.exit called'
-      );
-      expect(processExitSpy).toHaveBeenCalledWith(1);
-
-      axiosPostSpy.mockRestore();
-    });
   });
 });
