+++ src/types/extraction.test.ts
@@ -7,23 +7,16 @@ import {
   TimeValueType,
 } from './extraction';
 
-// Test the EventContext interface and related extraction types
 describe('ExtractionTypes', () => {
   const baseEvent = createMockEvent(mockServer.baseUrl, {
     payload: { event_type: EventType.StartExtractingData },
   });
 
-  it('should create event context without optional fields', () => {
-    const event = { ...baseEvent };
-
-    // If this compiles, the test passes
-    expect(event).toBeDefined();
-    expect(event.payload.event_context).toBeDefined();
-  });
-
   it('should create event context with all optional fields', () => {
+    // Arrange
     const event = { ...baseEvent };
 
+    // Act
     event.payload.event_context = {
       ...baseEvent.payload.event_context,
       extract_from: '2024-01-01T00:00:00Z',
@@ -32,7 +25,7 @@ describe('ExtractionTypes', () => {
       reset_extract_from: true,
     } as EventContext;
 
-    expect(event).toBeDefined();
+    // Assert
     expect(event.payload.event_context.extract_from).toBe(
       '2024-01-01T00:00:00Z'
     );
@@ -44,44 +37,54 @@ describe('ExtractionTypes', () => {
   });
 
   it('should create event context with partial optional fields', () => {
+    // Arrange
     const event = { ...baseEvent };
 
+    // Act
     event.payload.event_context = {
       ...baseEvent.payload.event_context,
       extract_from: '2024-01-01T00:00:00Z',
     } as EventContext;
 
-    expect(event).toBeDefined();
+    // Assert
     expect(event.payload.event_context.extract_from).toBe(
       '2024-01-01T00:00:00Z'
     );
   });
 
   it('should handle different InitialSyncScope values', () => {
+    // Arrange
     const event = { ...baseEvent };
 
+    // Act
     event.payload.event_context = {
       ...baseEvent.payload.event_context,
       initial_sync_scope: InitialSyncScope.FULL_HISTORY,
     } as EventContext;
 
+    // Assert
     expect(event.payload.event_context.initial_sync_scope).toBe(
       InitialSyncScope.FULL_HISTORY
     );
   });
 
   it('[edge] should handle null event context gracefully', () => {
+    // Arrange
     const event = { ...baseEvent };
 
+    // Act
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     event.payload.event_context = null as any;
 
+    // Assert
     expect(event.payload.event_context).toBeNull();
   });
 
   it('[edge] should handle undefined optional fields', () => {
+    // Arrange
     const event = { ...baseEvent };
 
+    // Act
     event.payload.event_context = {
       ...baseEvent.payload.event_context,
       extract_from: undefined,
@@ -90,6 +93,7 @@ describe('ExtractionTypes', () => {
       reset_extract_from: undefined,
     } as EventContext;
 
+    // Assert
     expect(event.payload.event_context.extract_from).toBeUndefined();
     expect(event.payload.event_context.extract_to).toBeUndefined();
     expect(event.payload.event_context.initial_sync_scope).toBeUndefined();
@@ -97,6 +101,7 @@ describe('ExtractionTypes', () => {
   });
 
   it('[edge] should handle explicit boolean values for reset_extract_from', () => {
+    // Arrange & Act
     const eventWithTrue = createMockEvent(mockServer.baseUrl, {
       payload: {
         event_type: EventType.StartExtractingData,
@@ -105,7 +110,6 @@ describe('ExtractionTypes', () => {
         },
       },
     });
-
     const eventWithFalse = createMockEvent(mockServer.baseUrl, {
       payload: {
         event_type: EventType.StartExtractingData,
@@ -115,6 +119,7 @@ describe('ExtractionTypes', () => {
       },
     });
 
+    // Assert
     expect(eventWithTrue.payload.event_context.reset_extract_from).toBe(true);
     expect(eventWithFalse.payload.event_context.reset_extract_from).toBe(false);
     expect(typeof eventWithTrue.payload.event_context.reset_extract_from).toBe(
@@ -127,6 +132,7 @@ describe('ExtractionTypes', () => {
 
   describe('TimeValueType enum', () => {
     it('should have all expected values', () => {
+      // Assert
       expect(TimeValueType.WORKERS_OLDEST).toBe('workers_oldest');
       expect(TimeValueType.WORKERS_OLDEST_MINUS_WINDOW).toBe(
         'workers_oldest_minus_window'
@@ -141,7 +147,10 @@ describe('ExtractionTypes', () => {
     });
 
     it('should have exactly seven values', () => {
+      // Act
       const values = Object.values(TimeValueType);
+
+      // Assert
       expect(values.length).toBe(7);
     });
   });
